+++ packages/expo-modules-core/ios/ReactDelegates/ExpoReactDelegate.swift
@@ -41,11 +41,4 @@ public class ExpoReactDelegate: NSObject {
       .compactMap { $0.bundleURL(reactDelegate: self) }
       .first(where: { _ in true })
   }
-
-  @objc
-  public func createRootViewController() -> UIViewController {
-    return self.handlers.lazy
-      .compactMap { $0.createRootViewController(reactDelegate: self) }
-      .first(where: { _ in true }) ?? UIViewController()
-  }
 }